+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deleting an object now clears the selection in tree views (previously selection would become an arbitrary object).
 - Fixed bug where adding/moving cohort sets between containers (INTERSECT/UNION/EXCEPT) could result in 2 objects with the same Order in the same container (resulting in ambiguous order of execution).
 - Fixed UI bug where selecting an extractable Catalogue would hide it's extractable (small green e) icon overlay
+- Fixed bug where deleting a Pinned object would not unpin the object
 
 ## [3.1.0]